Link epicctl statically
We're getting coredumps after integrating epicctl into a snapcraft
package. Let's see if linking it staticly will help.

Update: it turns out that we can work around the snapcraft problem,
but since the static epicctl exe is actually smaller than the dynamic
one we might as well stick with this.