Change the "create org" command to "create user-namespace"
Based on our work with the snap installer, what we used to call
organizations we now call user namespaces.

This also creates the Account CR in the namespace so it should be the
minimum needed.

I removed the uuidoverride flag because I can't remember what it was
good for, and user namespaces with UUIDs for names seems like a
user-hostile idea to me.