Use viper to read parameters
Cobra has command-line flags, but viper also reads from a
configuration file and environment variables. This is especially
important for commands that require an account and a service-group -
now those can be configured in a config file so they don't have to be
passed on the command line every single time.