Remove unused OwningClusterLabel
At some point we used the epicv1.OwningClusterLabel label to query for
a subset of remote endpoints, but that can cause a problem because the
cluster name can be longer than a k8s label is allowed to be.

It turns out that we no longer need this label so we can eliminate a
potential problem by not setting it.