Add a command to create ad-hoc linux endpoints
We want EPIC to be able to use ad-hoc linux backend nodes, i.e., nodes
that aren't part of k8s clusters. This is the first part of that
support. The create ad-hoc-endpoint command creates a GWEndpointSlice
CR on the epic cluster that allows GWRoutes to use the ad-hoc node as
an endpoint both in Envoy and TrueIngress.

The second part will be an agent that runs on the node and maintains
the TrueIngress tunnels. I'll probably build that in the
resource-model project since it may look somewhat like our existing
node agent.

Signed-off-by: Toby Cabot <toby@acnodal.io>